Add error for unsupported column types in `parse_column_type()`
Before this commit, the match expression in `parse_column_type()` did not have a default
case being applied. That meant that if the function was supplied with an unsupported column
type, an error without explanation would be thrown.

This commit adds the default case throwing a `ValueError` for unsupported column types.