Add type checking for `expression` in `SchemaColumnExpression`
Without proper type checking for `expression`, the user can provide an unsupported
data type to an expression in the schema file, which will then be passed to `SchemaColumnExpression`.

Although the code initially would not throw an error, Spark would eventually raise an exception, as
only strings are supported as expressions.

This commit adds type checking in `SchemaColumnExpression` and throws an error if it is anything
other than a string.