Enable test for parsing an unsupported column type
The test for checking unsupported column types in `parse_column_type()` was written in 83536705a7d2,
but was disabled since the function being tested did not throw an error in such cases.

The default case which throws an error for unsupported column types in `parse_column_type()` was introduced
in 1382357841d3 so we can enable this test now.