Fix issue with detecting boolean columns for Random Columns
This commit unblocks boolean types for Random columns from being processed by `parse_column_type()`. Booleans are automatically detected when both `min` and
`max` are not supplied in the schema. Although this was implemented correctly in `SchemaColumnRandomFactory`, when empty `min` and `max` values were passed
to `parse_column_type()` at the same type, it resulted in a Value Error.