Add unit and integration tests for Fixed and Expression-type columns
This commit adds unit tests for `SchemaColumnFixed` and `SchemaColumnExpression`, and both unit and integration tests for `parse_column_type()` for Fixed
and Expression-type columns.

For Fixed columns, I noted a design choice that I will leave to the future: currently (and in the original implementation in Scala), leaving the `value`
field empty for Fixed columns results in an error. However, that is nothing inherently wrong with passing something like `None` to Spark's `lit()`. It would
only result in a column of `null` values. I have added a TODO to return to this topic at a later date.