Add assertion for same-type list in Selection columns
Because Spark columns require a supported type, we choose the relevant type in `SchemaColumnSelection` by analyzing the type of the first value provided to
the `values` list in the schema file. However, lists of multiple different types are possible and this could result in an error down the line. This commit
adds a check to make sure all elements of the list are of the same type. This check happens both in `SchemaColumnSelection` and in `parse_column_type()`.

We also add a check in case `values` is not a list or is empty (but not missing).